% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_profanity.R
\name{check_profanity}
\alias{check_profanity}
\title{Check for profanity in a string (from One4All package)}
\usage{
check_profanity(x)
}
\arguments{
\item{x}{A character string to check for profanity.}
}
\value{
A logical value indicating whether the input string contains no profane words.
}
\description{
This function checks if the input string contains any profane words.
}
