% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{convert_pcm_to_mp3}
\alias{convert_pcm_to_mp3}
\title{Convert PCM audio to MP3 audio}
\usage{
convert_pcm_to_mp3(aud)
}
\arguments{
\item{aud}{Base64 decoded audio (pcm)}
}
\value{
Base64 encoded audio (mp3)
}
\description{
Convert PCM audio to MP3 audio
}
