% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{copy_template}
\alias{copy_template}
\title{Copy QMD Revealjs template file to a directory for use with \code{create_slides}}
\usage{
copy_template(dest_dir, audio = FALSE)
}
\arguments{
\item{dest_dir}{Path to directory where you want to copy the template file}

\item{audio}{\code{TRUE} will copy a template QMD that takes in audio inputs, while \code{FALSE} (default) will copy a template QMD that does not take audio inputs (only text and images)}
}
\value{
Copies a QMD file and provide the path to the destination directory
}
\description{
Copy QMD Revealjs template file to a directory for use with \code{create_slides}
}
