% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_image_prompts.R
\name{format_image_prompts}
\alias{format_image_prompts}
\title{Get prompt and negative prompt based on style}
\usage{
format_image_prompts(image_prompts, style)
}
\arguments{
\item{image_prompts}{Vector of prompts for generating images}

\item{style}{A character vector specifying the desired style. Choices are:
"anime", "comics", "lego_movie", "play_doh", "ethereal_fantasy", "isometric",
"line_art", "origami", "pixel_art", "abstract", "impressionist",
"renaissance", "watercolor", "biomechanical", "retro_futuristic",
"fighting_game", "mario", "pokemon", "street_fighter", "horror",
"manga", "space", "paper_mache", "tilt_shift".}
}
\value{
List of prompt and negative prompt
}
\description{
Get prompt and negative prompt based on style
}
