% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{generate_audio}
\alias{generate_audio}
\title{Get audio from story}
\usage{
generate_audio(story)
}
\arguments{
\item{text}{Vector of story text}
}
\value{
List of audio in binary.
}
\description{
Get audio from story
}
