% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_image_prompts.R
\name{generate_image_prompts}
\alias{generate_image_prompts}
\title{Generate Image Prompts using LLM}
\usage{
generate_image_prompts(
  story,
  system_prompt = NULL,
  chat_fn = ellmer::chat_google_gemini,
  model = "gemini-2.0-flash",
  ...
)
}
\arguments{
\item{story}{Character. Provide the story text generated by \code{generate_story}}

\item{system_prompt}{Character. System prompt for the client that generates prompts for creating images. Default system prompt is specified by an internal function that provides an example. See \code{storytelling:::generate_image_prompts_system_prompt}. If you provide it yourself, make sure you specify the model generates the same number of prompts as the length of story}

\item{chat_fn}{Ellmer client function. Default is \code{ellmer::chat_google_gemini}}

\item{model}{LLM model. Default is \code{"gemini-2.0-flash"}}

\item{...}{Additional named arguments that go in \code{chat_fn}. You may include the API key here using \code{api_key} argument, but it is recommended to provide the API key in your \code{.Renviron} file using \code{usethis::edit_r_environ()}. For Gemini, you may get an API key \href{https://ai.google.dev/gemini-api/docs/api-key}{here}. See \href{https://ellmer.tidyverse.org/}{ellmer docs} for more provider options}
}
\value{
Character vector containing prompts for images of length equal to \code{story}
}
\description{
Generate Image Prompts using LLM
}
\examples{
\dontrun{
library(storytelling)
user_prompt <- "Tell me a story about a dragon who turned into a human."
story <- generate_story(user_prompt)
image_prompts <- generate_image_prompts(story)
}
}
