% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_image_prompts_system_prompt.R
\name{generate_image_prompts_system_prompt}
\alias{generate_image_prompts_system_prompt}
\title{Generate the System Prompt for the Client that will write the prompts for generating images}
\usage{
generate_image_prompts_system_prompt(num_paras)
}
\arguments{
\item{num_paras}{Integer. Number of paragraphs in the story}
}
\value{
String containing system prompt
}
\description{
Generate the System Prompt for the Client that will write the prompts for generating images
}
