% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_images.R
\name{generate_images}
\alias{generate_images}
\title{Get all images corresponding to prompts}
\usage{
generate_images(
  image_prompts,
  style = c("anime", "comics", "lego_movie", "play_doh", "ethereal_fantasy", "line_art",
    "origami", "pixel_art", "impressionist", "watercolor", "biomechanical",
    "retro_futuristic", "fighting_game", "mario", "pokemon", "street_fighter", "horror",
    "manga", "space", "tilt_shift"),
  ...
)
}
\arguments{
\item{image_prompts}{A vector of prompts that describe what to draw}

\item{style}{A character vector specifying the desired style. Choices are:
"anime", "comics", "lego_movie", "play_doh", "ethereal_fantasy",
"line_art", "origami", "pixel_art", "impressionist",
"watercolor", "biomechanical", "retro_futuristic",
"fighting_game", "mario", "pokemon", "street_fighter", "horror",
"manga", "space", "tilt_shift".}

\item{...}{Additional named arguments provided to Stable Diffusion model API. See \href{https://developers.cloudflare.com/workers-ai/models/stable-diffusion-xl-base-1.0/}{stable-diffusion-xl-base-1.0 docs} for more details}
}
\value{
List of raw images.
}
\description{
Get all images corresponding to prompts
}
\details{
\code{generate_images} uses the internal function \code{storytelling:::req_single_image} to request an image from Stable Diffusion XL Base 1.0 model via Cloudflare API. Therefore, you need to create an account on Cloudflare and get an API key, and then set \code{CLOUDFLARE_ACCOUNT_ID} and \code{CLOUDFLARE_API_KEY} in your \code{.Renviron} file.
\href{https://developers.cloudflare.com/workers-ai/get-started/rest-api/}{Follow these instructions} to get the account ID and API key.
}
\examples{
\dontrun{
library(storytelling)
user_prompt <- "Tell me a story of a boy who learned to fly."
story <- generate_story(user_prompt)
image_prompts <- generate_image_prompts(story)
all_images <- generate_images(image_prompts, style = "ethereal_fantasy")
}
}
