% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_story.R
\name{generate_story}
\alias{generate_story}
\title{Generate a Fictional Story with LLM}
\usage{
generate_story(
  user_prompt,
  num_paras = 5,
  genre = c("adventure", "horror", "scifi", "mystery", "thriller", "fantasy", "comedy",
    "drama", "detective", "satire", "supernatural", "romance", "western"),
  struc = c("Hero's Journey", "Three Act Structure", "Seven-Point Story Structure",
    "Freytag's Pyramid", "Story Circle", "Save The Cat", "Fichtean Curve"),
  chat_fn = ellmer::chat_google_gemini,
  model = "gemini-2.0-flash",
  ...
)
}
\arguments{
\item{user_prompt}{Character. Prompt about story}

\item{num_paras}{Integer. Number of paragraphs in the generated story. Default is 5}

\item{genre}{Genre of the story. One of "adventure", "horror", "scifi", "mystery", "thriller",
"fantasy", "comedy", "drama", "detective", "satire", "supernatural", "romance" or "western"}

\item{struc}{Structure of the story. One of "Hero's Journey", "Three Act Structure", "Seven-Point Story Structure", "Freytag's Pyramid", "Story Circle", "Save The Cat" or "Fichtean Curve"}

\item{chat_fn}{Ellmer client function. Default is \code{ellmer::chat_google_gemini}}

\item{model}{LLM model. Default is \code{"gemini-2.0-flash"}}

\item{...}{Additional named arguments that go in \code{chat_fn}. You may include the API key here using \code{api_key} argument, but it is recommended to provide the API key in your \code{.Renviron} file using \code{usethis::edit_r_environ()}. For Gemini, you may get an API key \href{https://ai.google.dev/gemini-api/docs/api-key}{here}. See \href{https://ellmer.tidyverse.org/}{ellmer docs} for more provider options}
}
\value{
Character vector containing story text of length equal to \code{num_paras}
}
\description{
Generate a Fictional Story with LLM
}
\examples{
\dontrun{
library(storytelling)
user_prompt <- "Tell me a story of a boy who learned to fly."
generate_story(user_prompt)
generate_story(user_prompt, genre = "thriller", struc = "Freytag's Pyramid")
}
}
