% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_story_system_prompt.R
\name{generate_story_system_prompt}
\alias{generate_story_system_prompt}
\title{Generate System Prompt for Client that generates Stories}
\usage{
generate_story_system_prompt(genre, struc, num_paras = 5)
}
\arguments{
\item{genre}{Genre of the story}

\item{struc}{Structure of the story}

\item{num_paras}{Number of paragraphs in the story}
}
\value{
System prompt as a string
}
\description{
Generate System Prompt for Client that generates Stories
}
