% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{get_raw_audio}
\alias{get_raw_audio}
\title{Extract Raw Audio}
\usage{
get_raw_audio(res, model = c("cloudflare", "google"))
}
\arguments{
\item{res}{API response}

\item{model}{Text to speech model. Options are "google" and "cloudflare"}
}
\value{
Audio in binary.
}
\description{
Extract Raw Audio
}
