% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chat.R
\name{make_chat}
\alias{make_chat}
\title{Make an ellmer chat object}
\usage{
make_chat(chat_fn, ...)
}
\arguments{
\item{chat_fn}{Ellmer chat function, e.g., \code{chat_google_gemini}}

\item{...}{Parameters of ellmer chat function}
}
\value{
Ellmer chat object
}
\description{
Make an ellmer chat object
}
