% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_single_image.R
\name{req_single_image}
\alias{req_single_image}
\title{Request a single image from Stable Diffusion XL Base 1.0 model via Cloudflare API}
\usage{
req_single_image(
  prompt,
  negative_prompt,
  num_steps = 20,
  ACCOUNT_ID = Sys.getenv("CLOUDFLARE_ACCOUNT_ID"),
  API_KEY = Sys.getenv("CLOUDFLARE_API_KEY"),
  ...
)
}
\arguments{
\item{prompt}{Description of image}

\item{negative_prompt}{Description of what to exclude}

\item{num_steps}{Number of diffusion steps. Max 20}

\item{ACCOUNT_ID}{Cloudflare Workers AI Model API account ID}

\item{API_KEY}{Cloudflare Workers AI Model API key}

\item{...}{Additional named arguments provided to Stable Diffusion model API. See \href{https://developers.cloudflare.com/workers-ai/models/stable-diffusion-xl-base-1.0/}{stable-diffusion-xl-base-1.0 docs} for more details}
}
\value{
Request.
}
\description{
Request a single image from Stable Diffusion XL Base 1.0 model via Cloudflare API
}
