% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{generate_audio}
\alias{generate_audio}
\title{Get audio from story}
\usage{
generate_audio(story, voice = NULL, model = c("cloudflare", "google"))
}
\arguments{
\item{story}{Vector of story text}

\item{voice}{Google text-to-speech voice name. Not compatible with \code{model = "cloudflare"}. One of 36 voices can be specified. See Details.}

\item{model}{Text to speech model. Options are "google" and "cloudflare"}
}
\value{
List of audio in binary.
}
\description{
Get audio from story
}
\details{
Visit \href{https://ai.google.dev/gemini-api/docs/speech-generation#voices}{speech generation voice options}
for details about the 36 voice options.
}
\examples{
\dontrun{
library(storytelling)
user_prompt <- "Tell me a story about a dragon who turned into a human."
story <- generate_story(user_prompt)
audios <- generate_audio(story)
}
}
