% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{generate_audio}
\alias{generate_audio}
\title{Get audio from story}
\usage{
generate_audio(story, model = c("cloudflare", "google"))
}
\arguments{
\item{story}{Vector of story text}

\item{model}{Text to speech model. Options are "google" and "cloudflare"}
}
\value{
List of audio in binary.
}
\description{
Get audio from story
}
\examples{
\dontrun{
library(storytelling)
user_prompt <- "Tell me a story about a dragon who turned into a human."
story <- generate_story(user_prompt)
audios <- generate_audio(story)
}
}
