% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{get_audio_req}
\alias{get_audio_req}
\title{Convert text to speech}
\usage{
get_audio_req(
  text,
  voice = NULL,
  model = c("cloudflare", "google"),
  CF_ACCOUNT_ID = Sys.getenv("CLOUDFLARE_ACCOUNT_ID"),
  CF_API_KEY = Sys.getenv("CLOUDFLARE_API_KEY"),
  G_API_KEY = Sys.getenv("GOOGLE_API_KEY")
)
}
\arguments{
\item{text}{Text to be converted to speech}

\item{voice}{Google text-to-speech voice name. Not compatible with \code{model = "cloudflare"}. One of 36 voices can be specified.}

\item{model}{Text to speech model. Options are "google" and "cloudflare"}

\item{CF_ACCOUNT_ID}{Cloudflare Workers AI Model API account ID}

\item{CF_API_KEY}{Cloudflare Workers AI Model API key}

\item{G_API_KEY}{Google API key}
}
\value{
httr2 request.
}
\description{
Convert text to speech
}
